+++ CHANGES.rst
@@ -16,6 +16,9 @@ Unreleased
 - Show additional information about the scaling capabilities of a product in
   the output of the ``croud products list`` command.
 
+0.15.2 - 2019/07/09
+===================
+
 - Deprecated ``croud users roles add`` and ``croud users roles remove`` in
   favor of project and organization level commands ``croud
   organizations|projects users add|remove``.